+++ src/Components/Message/index.js
@@ -1,15 +1,6 @@
-import React from 'react'
-import useRightClickMenu from "../../hooks/useRightClickMenu";
-import ContextMenu from '../ContextMenu';
-
 const Message = () => {
-    const {x, y, showMenu} = useRightClickMenu();
     return (
-
-        <div className="overflow-y-scroll h-4/5">
-
-            <ContextMenu x={x} y={y} showMenu={showMenu} />
-
+        <>
             <div className="flex justify-end px-10 pt-5">
                 <div className="mx-3">
                     <div className="text-right">
@@ -45,8 +36,8 @@ const Message = () => {
                     </div>
                 </div>
             </div>
-        </div>
-        
+
+        </>
     )
 }
