+++ src/Components/Setting/index.js
@@ -1,5 +1,4 @@
-import React from 'react'
-import profile from './alex-suprun-1JGHAAdbL_Y-unsplash.jpg'
+import profile from '../../images/avatar-profile.jpg'
 
 const Setting = ({ setShowSetting, closeSidebar }) => {
 
@@ -94,7 +93,7 @@ const Setting = ({ setShowSetting, closeSidebar }) => {
           <li className="flex items-center py-3.5 hover:bg-gray-200 dark:hover:bg-slate-700 hover:pr-3 ease-in-out duration-300 cursor-pointer rounded-2xl">
             <svg xmlns="http://www.w3.org/2000/svg" fill="currentColor" className="w-5 h-5 text-red-600" viewBox="0 0 16 16">
               <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6z" />
-              <path fill-rule="evenodd" d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z" />
+              <path fillRule="evenodd" d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z" />
             </svg>
             <p className='pr-3'>
               خروج از حساب
