+++ src/Pages/Login/index.js
@@ -1,0 +1,7 @@
+const Login = () => {
+  return (
+    <div>Login</div>
+  )
+}
+
+export default Login
